move comment test into test_counter
